Local integration test is working with Package, Validate and Store.

+++ dpn/validator.go
@@ -61,7 +61,7 @@ func (validator *Validator) HandleMessage(message *nsq.Message) error {
 	validator.ValidationChannel <- dpnResult
 	// identifier or dpn identifier
 	validator.ProcUtil.MessageLog.Info("Put %s into validation channel",
-		dpnResult.BagIdentifier)
+		dpnResult.DPNBag.UUID)
 	return nil
 }
 
@@ -69,6 +69,12 @@ func (validator *Validator) validate() {
 	for result := range validator.ValidationChannel {
 		if result.NsqMessage != nil {
 			result.NsqMessage.Touch()
+		}
+		if result.LocalPath == "" {
+			result.ErrorMessage = "Cannot validate bag because DPNResult.LocalPath is not set. " +
+				"This should be set to the location of the tar file you want to validate."
+			validator.PostProcessChannel <- result
+			continue
 		}
 		var err error
 		// Set up a proper validation result object for this bag.
@@ -78,6 +84,7 @@ func (validator *Validator) validate() {
 				"Could not create ValidationResult for bag %s: %v",
 				result.DPNBag.UUID, err)
 			validator.PostProcessChannel <- result
+			continue
 		}
 		// Now validate the bag. This step can take a long time on
 		// large bags, since we may be untarring hundred of gigabytes
@@ -142,12 +149,19 @@ func (validator *Validator) postProcess() {
 			result.NsqMessage.Touch()
 		}
 		if result.ErrorMessage != "" {
+			validator.ProcUtil.MessageLog.Error(result.ErrorMessage)
 			validator.ProcUtil.IncrementFailed()
 			SendToTroubleQueue(result, validator.ProcUtil)
 			if bagman.FileExists(result.ValidationResult.TarFilePath) {
 				os.Remove(result.ValidationResult.TarFilePath)
+				validator.ProcUtil.MessageLog.Debug(
+					"Deleting tar file %s", result.ValidationResult.TarFilePath)
 			}
-			if bagman.FileExists(result.ValidationResult.UntarredPath) {
+			if result.ValidationResult.UntarredPath != "" &&
+				result.ValidationResult.UntarredPath != "/" &&
+				bagman.FileExists(result.ValidationResult.UntarredPath) {
+				validator.ProcUtil.MessageLog.Debug(
+					"Deleting directory %s and its contents", result.ValidationResult.UntarredPath)
 				os.RemoveAll(result.ValidationResult.UntarredPath)
 			}
 		} else {
@@ -210,31 +224,54 @@ func (validator *Validator) updateRemoteNode(result *DPNResult) {
 		validator.DPNConfig.RestClient.LocalAPIRoot, // All nodes should be on same version as local
 		authToken,
 		validator.ProcUtil.MessageLog)
+	if err != nil {
+		result.ErrorMessage = fmt.Sprintf("Could not create REST client for remote node %s: %v",
+			remoteNode.Namespace, err)
+		return
+	}
+
 
 	// Update the transfer request and send it back to the remote node.
 	// We'll get an updated transfer request back from that node.
+	bagValid := result.ValidationResult.IsValid()
 	result.TransferRequest.Status = "Received"
+	result.TransferRequest.BagValid = &bagValid
 	result.TransferRequest.FixityValue = result.ValidationResult.TagManifestChecksum
+
+	validator.ProcUtil.MessageLog.Debug("Updating xfer request %s status for bag %s on remote node %s. " +
+		"Setting status to 'Received', BagValid to %t, and tag manifest checksum to %s",
+		result.TransferRequest.ReplicationId, result.TransferRequest.UUID,
+		remoteNode.Namespace, *result.TransferRequest.BagValid,
+		result.TransferRequest.FixityValue)
 	xfer, err := remoteRESTClient.ReplicationTransferUpdate(result.TransferRequest)
 	if err != nil {
 		result.ErrorMessage = fmt.Sprintf("Error updating transfer request on remote node: %v", err)
 		return
 	}
 	if *xfer.FixityAccept == false {
+		validator.ProcUtil.MessageLog.Debug(
+			"Remote node rejected fixity value for xfer request %s (bag %s)",
+			result.TransferRequest.ReplicationId, result.TransferRequest.UUID)
 		result.ErrorMessage = "Remote node did not accept the fixity value we sent for this bag. " +
 			"This cancels the transfer request, and we will not store the bag."
 		return
 	}
 	if xfer.Status == "Cancelled" {
+		validator.ProcUtil.MessageLog.Debug(
+			"Remote node says status is 'Cancelled' for xfer request %s (bag %s)",
+			result.TransferRequest.ReplicationId, result.TransferRequest.UUID)
 		result.ErrorMessage = "This transfer request has been marked as cancelled on the remote node. " +
 			"This bag will not be copied to storage."
+		return
 	}
+	validator.ProcUtil.MessageLog.Debug("Remote node updated xfer request %s (bag %s), " +
+		"and set status to %s", xfer.ReplicationId, xfer.UUID, xfer.Status)
 }
 
 func (validator *Validator) RunTest(result *DPNResult) {
 	validator.WaitGroup.Add(1)
 	validator.ProcUtil.MessageLog.Info("Putting %s into validation channel",
-		result.BagIdentifier)
+		result.DPNBag.UUID)
 	validator.ValidationChannel <- result
 	validator.WaitGroup.Wait()
 	if result.ErrorMessage != "" {
